Move SkillDirector and WebhookDirector
