+++ glyph-bot/src/main/kotlin/Glyph.kt
@@ -36,22 +36,15 @@ import org.yttr.glyph.bot.directors.StatusDirector
 import org.yttr.glyph.bot.directors.config.ConfigDirector
 import org.yttr.glyph.bot.directors.servers.BotList
 import org.yttr.glyph.bot.directors.servers.ServerDirector
-import org.yttr.glyph.bot.directors.skills.SkillDirector
 import org.yttr.glyph.bot.directors.starboard.StarboardDirector
 import org.yttr.glyph.bot.messaging.MessagingDirector
 import org.yttr.glyph.bot.messaging.quickview.QuickviewDirector
-import org.yttr.glyph.bot.skills.DoomsdayClockSkill
-import org.yttr.glyph.bot.skills.EphemeralSaySkill
 import org.yttr.glyph.bot.skills.FallbackSkill
-import org.yttr.glyph.bot.skills.FeedbackSkill
-import org.yttr.glyph.bot.skills.HelpSkill
-import org.yttr.glyph.bot.skills.RankSkill
-import org.yttr.glyph.bot.skills.RedditSkill
-import org.yttr.glyph.bot.skills.ServerConfigSkill
-import org.yttr.glyph.bot.skills.SnowstampSkill
-import org.yttr.glyph.bot.skills.SourceSkill
-import org.yttr.glyph.bot.skills.StatusSkill
-import org.yttr.glyph.bot.skills.TimeSkill
+import org.yttr.glyph.bot.skills.SkillDirector
+import org.yttr.glyph.bot.skills.`fun`.DoomsdayClockSkill
+import org.yttr.glyph.bot.skills.`fun`.EphemeralSaySkill
+import org.yttr.glyph.bot.skills.`fun`.RankSkill
+import org.yttr.glyph.bot.skills.`fun`.RedditSkill
 import org.yttr.glyph.bot.skills.creator.ChangeStatusSkill
 import org.yttr.glyph.bot.skills.moderation.BanSkill
 import org.yttr.glyph.bot.skills.moderation.GuildInfoSkill
@@ -61,6 +54,13 @@ import org.yttr.glyph.bot.skills.moderation.UserInfoSkill
 import org.yttr.glyph.bot.skills.roles.RoleListSkill
 import org.yttr.glyph.bot.skills.roles.RoleSetSkill
 import org.yttr.glyph.bot.skills.roles.RoleUnsetSkill
+import org.yttr.glyph.bot.skills.util.FeedbackSkill
+import org.yttr.glyph.bot.skills.util.HelpSkill
+import org.yttr.glyph.bot.skills.util.ServerConfigSkill
+import org.yttr.glyph.bot.skills.util.SnowstampSkill
+import org.yttr.glyph.bot.skills.util.SourceSkill
+import org.yttr.glyph.bot.skills.util.StatusSkill
+import org.yttr.glyph.bot.skills.util.TimeSkill
 import org.yttr.glyph.bot.skills.wiki.WikiSkill
 import org.yttr.glyph.shared.pubsub.redis.RedisAsync
