+++ glyph-bot/src/main/kotlin/skills/fun/RedditSkill.kt
@@ -4,7 +4,7 @@
  * Glyph, a Discord bot that uses natural language instead of commands
  * powered by DialogFlow and Kotlin
  *
- * Copyright (C) 2017-2020 by Ian Moore
+ * Copyright (C) 2017-2021 by Ian Moore
  *
  * This file is part of Glyph.
  *
@@ -20,9 +20,7 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.yttr.glyph.bot.skills
+ */kage org.yttr.glyph.bot.skills.`fun`
 
 import com.squareup.moshi.JsonDataException
 import net.dean.jraw.ApiException
@@ -41,8 +39,8 @@ import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import net.jodah.expiringmap.ExpiringMap
 import org.yttr.glyph.bot.Glyph
 import org.yttr.glyph.bot.ai.AIResponse
-import org.yttr.glyph.bot.directors.skills.Skill
 import org.yttr.glyph.bot.messaging.Response
+import org.yttr.glyph.bot.skills.Skill
 import java.time.Instant
 import java.util.LinkedList
 import java.util.Queue
