+++ glyph-bot/src/main/kotlin/skills/util/SourceSkill.kt
@@ -1,10 +1,10 @@
 /*
- * ServerConfigSkill.kt
+ * SourceSkill.kt
  *
  * Glyph, a Discord bot that uses natural language instead of commands
  * powered by DialogFlow and Kotlin
  *
- * Copyright (C) 2017-2020 by Ian Moore
+ * Copyright (C) 2017-2021 by Ian Moore
  *
  * This file is part of Glyph.
  *
@@ -20,26 +20,31 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
+ */kage org.yttr.glyph.bot.skills.util
 
-package org.yttr.glyph.bot.skills
-
-import net.dv8tion.jda.api.Permission
+import net.dv8tion.jda.api.EmbedBuilder
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent
+import org.yttr.glyph.bot.Glyph
 import org.yttr.glyph.bot.ai.AIResponse
-import org.yttr.glyph.bot.directors.skills.Skill
 import org.yttr.glyph.bot.messaging.Response
+import org.yttr.glyph.bot.skills.Skill
+import java.awt.Color
+import java.time.Instant
 
 /**
- * Tells people to use the config website to edit their config
+ * A skill that allows users to see the license and link to the source code
  */
-class ServerConfigSkill : Skill(
-    "skill.config.server",
-    cooldownTime = 15,
-    guildOnly = true,
-    requiredPermissionsSelf = listOf(Permission.MANAGE_WEBHOOKS),
-    requiredPermissionsUser = listOf(Permission.ADMINISTRATOR)
-) {
-    override suspend fun onTrigger(event: MessageReceivedEvent, ai: AIResponse): Response =
-        Response.Volatile("To edit your config, visit https://gl.yttr.org/config")
+class SourceSkill : Skill("skill.source") {
+    override suspend fun onTrigger(event: MessageReceivedEvent, ai: AIResponse): Response {
+        val name = event.jda.selfUser.name
+        val embed = EmbedBuilder()
+            .setTitle("$name Source")
+            .setDescription(ai.result.fulfillment.speech)
+            .setFooter("$name-Kotlin-${Glyph.version}", null)
+            .setTimestamp(Instant.now())
+            .setColor(Color.getHSBColor(0.6f, 0.89f, 0.61f))
+            .build()
+
+        return Response.Volatile(embed)
+    }
 }
